+++ server/src/main/scala/com/cloudera/livy/server/Main.scala
@@ -32,6 +32,7 @@ import com.cloudera.livy._
 import com.cloudera.livy.server.batch.BatchSessionServlet
 import com.cloudera.livy.server.client.ClientSessionServlet
 import com.cloudera.livy.server.interactive.InteractiveSessionServlet
+import com.cloudera.livy.util.LineBufferedProcess
 
 object Main extends Logging {
