+++ server/src/main/scala/com/cloudera/livy/sessions/Session.scala
