+++ server/src/main/scala/com/cloudera/livy/sessions/SessionManager.scala
