+++ server/src/test/scala/com/cloudera/livy/sessions/SessionManagerSpec.scala
