--- server/src/test/scala/com/cloudera/livy/spark/SparkProcessBuilderSpec.scala
@@ -1,44 +1,0 @@
-/*
- * Licensed to Cloudera, Inc. under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  Cloudera, Inc. licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.cloudera.livy.spark
-
-import org.scalatest.{FlatSpec, Matchers}
-
-import com.cloudera.livy.LivyConf
-
-class SparkProcessBuilderSpec
-  extends FlatSpec
-  with Matchers {
-
-  it should "not error if passed an allowed config option" in {
-    val allowed = Set("spark.app.name")
-    val builder = new SparkProcessBuilder(new LivyConf(), allowed)
-
-    builder.conf("spark.app.name", "hello")
-  }
-
-  it should "raise an exception if passed an illegal config option" in {
-    val allowed = Set("spark.app.name")
-    val builder = new SparkProcessBuilder(new LivyConf(), allowed)
-
-    intercept[ConfigOptionNotAllowed] {
-      builder.conf("spark.master", "local")
-    }
-  }
-}
